split crafting into own mod
