+++ crates/client/src/app/input/mod.rs
@@ -1,2 +1,3 @@
+pub mod crafting;
 pub mod hotbar_assignment;
 pub mod inventory;
